services/backend/requirements.txt
fastapi==0.79.0
uvicorn==0.16.0
aerich==0.6.3
asyncpg==0.26.0
tortoise-orm==0.19.2
bcrypt==3.2.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
